refactor(experimental): rename files and combineCodec function
