+++ packages/codecs-core/src/index.ts
@@ -1,6 +1,6 @@
 export * from './bytes';
 export * from './codec';
 export * from './errors';
-export * from './fixCodec';
-export * from './mapCodec';
-export * from './reverseCodec';
+export * from './fix-codec';
+export * from './map-codec';
+export * from './reverse-codec';
